Add failing test for view all for discography
